further improvement of the codebase with moving the persistence logic to the persistence layer
